+++ android/src/main/kotlin/co/tamara/sdk/log/Logging.kt
@@ -1,19 +1,19 @@
 package co.tamara.sdk.log
 
 import android.util.Log
-//import co.tamara.sdk.BuildConfig
+import co.tamara.sdk.BuildConfig
 
 internal class Logging {
     companion object{
-        //fun d(tag: String, message: String){
-        //    if(BuildConfig.DEBUG) {
-        //        Log.d(tag, message)
-        //    }
-        //}
-        //fun w(tag: String, message: String){
-        //    if(BuildConfig.DEBUG){
-        //        Log.w(tag, message)
-        //    }
-        //}
+        fun d(tag: String, message: String){
+            if(BuildConfig.DEBUG) {
+                Log.d(tag, message)
+            }
+        }
+        fun w(tag: String, message: String){
+            if(BuildConfig.DEBUG){
+                Log.w(tag, message)
+            }
+        }
     }
 }