+++ android/src/main/kotlin/co/tamara/sdk/ui/TamaraPaymentViewModel.kt
@@ -1,21 +1,15 @@
 package co.tamara.sdk.ui
 
-import android.app.Activity
 import androidx.lifecycle.*
 import co.tamara.sdk.DIHelper
-import co.tamara.sdk.PaymentResult
-import co.tamara.sdk.TamaraPayment
-import co.tamara.sdk.TamaraPaymentActivity
-import co.tamara.sdk.error.PaymentError
-import co.tamara.sdk.log.Logging
 import co.tamara.sdk.model.Order
 import co.tamara.sdk.model.response.CheckoutSession
 import co.tamara.sdk.model.response.PaymentType
 import co.tamara.sdk.repository.CheckOutRepository
+import co.tamara.sdk.util.SingleLiveEvent
 import co.tamara.sdk.vo.Resource
 import co.tamara.sdk.vo.Status
 import javax.inject.Inject
-
 
 internal class TamaraPaymentViewModel : ViewModel() {
     init {
@@ -25,9 +19,9 @@ internal class TamaraPaymentViewModel : ViewModel() {
     @Inject
     lateinit var repository: CheckOutRepository
 
-    private var orderLiveData = MutableLiveData<Order>()
+    private var orderLiveData = SingleLiveEvent<Order>()
 
-    var orderInfoLiveData: LiveData<Resource<CheckoutSession>> = Transformations.switchMap(orderLiveData){
+    var orderInfoLiveData: LiveData<Resource<CheckoutSession>> = orderLiveData.switchMap{
         repository.createOrder(it)
     }