+++ CHANGELOG.md
@@ -1,14 +1,47 @@
-## 1.0.0
-* initial  Version
+## 0.9.0
 
-## 1.0.1
-* gradle-8.3
+* Tamara Checkout package 
 
-## 1.0.2
-* ext.kotlin_version = '1.8.21'
+## 0.9.1
 
-## 1.0.3
-* Bug Fixes
+* Fix canceled callback
 
-## 1.0.4
-* SDK Update
+## 1.0.6
+
+* Update Flutter SDK with other features
+
+## 1.0.7 
+
+* Add instalment option
+
+## 1.0.8
+
+* Add Camera permission
+
+## 1.0.9
+
+* Fix issue with iOS for camera permission
+
+## 1.0.10
+
+* Fix issue with intl
+
+## 1.0.11
+
+* Update version
+
+## 1.0.12
+
+* Update version
+
+## 1.0.13
+
+* Update Library to support latest version
+
+## 1.0.14
+
+* Fix issue with handle events (success, cancel, failure)
+
+## 1.0.15
+
+* Update with webview